+++ spec/cask/audit_spec.rb
@@ -3,20 +3,20 @@ require 'spec_helper'
 describe Hbc::Audit do
   describe "result" do
     it "is 'failed' if there are have been any errors added" do
-      audit = Hbc::Audit.new(Hbc.new)
+      audit = Hbc::Audit.new(Hbc::Cask.new)
       audit.add_error 'bad'
       audit.add_warning 'eh'
       expect(audit.result).to match(/failed/)
     end
 
     it "is 'warning' if there are no errors, but there are warnings" do
-      audit = Hbc::Audit.new(Hbc.new)
+      audit = Hbc::Audit.new(Hbc::Cask.new)
       audit.add_warning 'eh'
       expect(audit.result).to match(/warning/)
     end
 
     it "is 'passed' if there are no errors or warning" do
-      audit = Hbc::Audit.new(Hbc.new)
+      audit = Hbc::Audit.new(Hbc::Cask.new)
       expect(audit.result).to match(/passed/)
     end
   end
@@ -73,7 +73,7 @@ describe Hbc::Audit do
         download = double()
         download.expects(:perform).raises(StandardError.new(error_message))
 
-        audit = Hbc::Audit.new(Hbc.new)
+        audit = Hbc::Audit.new(Hbc::Cask.new)
         audit.run!(download)
         expect(audit.errors).to include(/#{error_message}/)
       end
