+++ spec/cask/download_strategy_spec.rb
@@ -4,7 +4,7 @@ describe 'download strategies' do
   let(:url) { 'http://example.com/cask.dmg' }
   let(:url_options) { Hash.new }
   let(:cask) {
-    class_double(Hbc,
+    class_double(Hbc::Cask,
       :token => 'some-cask',
       :url => Hbc::URL.new(url, url_options),
       :version => '1.2.3.4'
