+++ test/cask/artifact/two_apps_correct_test.rb
@@ -20,7 +20,7 @@ describe Hbc::Artifact::App do
     end
 
     it "works with an application in a subdir" do
-      AltSubDirTwoAppsCask = Class.new(Hbc)
+      AltSubDirTwoAppsCask = Class.new(Hbc::Cask)
       AltSubDirTwoAppsCask.class_eval do
         url TestHelper.local_binary_url('caffeine.zip')
         homepage 'http://example.com/local-caffeine'
